+++ src/frame/types.ts
@@ -1,3 +1,4 @@
+import type { Abi, Address, Hex } from 'viem';
 import { NeynarFrameValidationInternalModel } from '../utils/neynar/frame/types';
 
 /**
@@ -15,7 +16,9 @@ export interface FrameData {
   fid: number;
   messageHash: string;
   network: number;
+  state: string;
   timestamp: number;
+  transactionId?: string;
   url: string;
 }
 
@@ -50,6 +53,9 @@ export interface FrameValidationData {
   liked: boolean; // Indicates if the viewer clicking the frame liked the cast
   raw: NeynarFrameValidationInternalModel;
   recasted: boolean; // Indicates if the viewer clicking the frame recasted the cast
+  state: {
+    serialized: string; // Serialized state (e.g. JSON) passed to the frame server
+  };
   valid: boolean; // Indicates if the frame is valid
 }
 
@@ -77,7 +83,7 @@ export function convertToFrame(json: any) {
  */
 export type FrameButtonMetadata =
   | {
-      action: 'link' | 'mint';
+      action: 'link' | 'mint' | 'tx';
       label: string;
       target: string;
     }
@@ -115,28 +121,43 @@ export type FrameMetadataReact = FrameMetadataType & {
  * Note: exported as public Type
  */
 export type FrameMetadataType = {
-  // A list of strings which are the label for the buttons in the frame (max 4 buttons).
-  buttons?: [FrameButtonMetadata, ...FrameButtonMetadata[]];
-  // An image which must be smaller than 10MB and should have an aspect ratio of 1.91:1
-  image: string | FrameImageMetadata;
-  // The text input to use for the Frame.
-  input?: FrameInputMetadata;
+  buttons?: [FrameButtonMetadata, ...FrameButtonMetadata[]]; // A list of strings which are the label for the buttons in the frame (max 4 buttons).
+  image: string | FrameImageMetadata; // An image which must be smaller than 10MB and should have an aspect ratio of 1.91:1
+  input?: FrameInputMetadata; // The text input to use for the Frame.
   /** @deprecated Prefer `postUrl` */
   post_url?: string;
-  // A valid POST URL to send the Signature Packet to.
-  postUrl?: string;
+  postUrl?: string; // A valid POST URL to send the Signature Packet to.
   /** @deprecated Prefer `refreshPeriod` */
   refresh_period?: number;
-  // A period in seconds at which the app should expect the image to update.
-  refreshPeriod?: number;
-  // A string containing serialized state (e.g. JSON) passed to the frame server.
-  state?: object;
+  refreshPeriod?: number; // A period in seconds at which the app should expect the image to update.
+  state?: object; // A string containing serialized state (e.g. JSON) passed to the frame server.
 };
 
 /**
  * Note: exported as public Type
  */
 export type FrameMetadataResponse = Record<string, string>;
+
+/**
+ * Note: exported as public Type
+ */
+type ChainNamespace = 'eip155' | 'solana';
+type ChainReference = string;
+export type FrameTransactionResponse = {
+  chainId: `${ChainNamespace}:${ChainReference}`; // A CAIP-2 chain ID to identify the tx network
+  method: 'eth_sendTransaction'; // A method ID to identify the type of tx request.
+  params: FrameTransactionEthSendParams; // Specific parameters for chainId and method
+};
+
+/**
+ * Note: exported as public Type
+ */
+export type FrameTransactionEthSendParams = {
+  abi: Abi; // The contract ABI for the contract to call.
+  data?: Hex; // The data to send with the transaction.
+  to: Address; // The address of the contract to call.
+  value: string; // The amount of Wei to send with the transaction
+};
 
 /**
  * Settings to simulate statuses on mock frames.